+++ go.mod
@@ -1,0 +1,11 @@
+module github.com/kristofmartens/mproxy
+
+go 1.14
+
+require (
+	github.com/dgrijalva/jwt-go v3.2.0+incompatible
+	github.com/go-yaml/yaml v2.1.0+incompatible
+	github.com/lestrrat/go-jwx v0.0.0-20180221005942-b7d4802280ae
+	github.com/lestrrat/go-pdebug v0.0.0-20180220043741-569c97477ae8 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+)
